backend/requirements.txt
# Django Backend
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-autoslug==1.9.9
whitenoise==6.8.2

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Task Queue
celery==5.4.0
redis==5.2.1

# Payments
stripe==11.4.1

# Images
cloudinary==1.41.0

# Email
resend==2.5.1

# Utilities
Pillow==11.1.0
python-dateutil==2.9.0

# Production
gunicorn==23.0.0

# Development
pytest==8.3.4
pytest-django==4.9.0
